+++ ItunesMedia/ViewControllers/MediaFeedViewController/MediaFeedViewController.swift
@@ -1,0 +1,20 @@
+//
+//  MediaFeedViewController.swift
+//  ItunesMedia
+//
+//  Created by Gustavo Lopez on 12/26/19.
+//  Copyright © 2019 Gustavo Lopez. All rights reserved.
+//
+
+import UIKit
+
+final class MediaFeedViewController: UIViewController {
+
+    // MARK: - LifeCycle
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .red
+    }
+
+}